refactor(plugin): extract main service helpers
