+++ packages/obsidian-plugin/src/main.ts
@@ -1,12 +1,10 @@
 import { Plugin } from "obsidian";
 
 import { registerCommands } from "./commands/registerCommands.js";
-import { AutoIndexScheduler } from "./indexer/autoIndexScheduler.js";
 import { registerAutoIndexEvents } from "./indexer/autoIndexEvents.js";
 import { saveLastIndexerLogToFile as saveLastIndexerLogToFileInVault } from "./indexer/indexerLogFile.js";
 import { getIndexerFailureHint } from "./indexer/indexerFailureHints.js";
-import { IndexerRunner, type AilssIndexerStatusSnapshot } from "./indexer/indexerRunner.js";
-import { McpHttpServiceController } from "./mcp/mcpHttpServiceController.js";
+import { type AilssIndexerStatusSnapshot } from "./indexer/indexerRunner.js";
 import type { AilssMcpHttpServiceStatusSnapshot } from "./mcp/mcpHttpServiceTypes.js";
 import { normalizeAilssPluginDataV1, parseAilssPluginData } from "./persistence/pluginData.js";
 import {
@@ -15,8 +13,17 @@ import {
 	runPythonApiAgent,
 	runPythonApiEval,
 } from "./pythonApi/client.js";
-import { PythonApiServiceController } from "./pythonApi/pythonApiServiceController.js";
+import {
+	formatPythonAgentResult,
+	formatPythonRetrieveResult,
+} from "./pythonApi/resultFormatting.js";
 import type { AilssPythonApiServiceStatusSnapshot } from "./pythonApi/pythonApiServiceTypes.js";
+import {
+	createAutoIndexScheduler,
+	createIndexerRunner,
+	createMcpHttpServiceController,
+	createPythonApiServiceController,
+} from "./runtime/pluginServiceFactories.js";
 import { startConfiguredServices } from "./runtime/startConfiguredServices.js";
 import {
 	AilssObsidianSettingTab,
@@ -53,13 +60,7 @@ import {
 } from "./utils/codexClipboardService.js";
 import { formatAilssTimestampForUi } from "./utils/dateTime.js";
 import { generateToken } from "./utils/misc.js";
-import {
-	getPluginDirRealpathOrNull,
-	getVaultPath,
-	resolveIndexerArgs,
-	resolveMcpHttpArgs,
-	resolvePythonApiArgs,
-} from "./utils/pluginPaths.js";
+import { getVaultPath } from "./utils/pluginPaths.js";
 import { type PromptKind } from "./utils/promptTemplates.js";
 import { installVaultRootPromptAtVaultRoot } from "./utils/vaultRootPromptInstaller.js";
 
@@ -74,70 +75,58 @@ export default class AilssObsidianPlugin extends Plugin {
 	private statusBarEl: HTMLElement | null = null;
 	private mcpStatusBarEl: HTMLElement | null = null;
 
-	private readonly mcpHttpService = new McpHttpServiceController({
+	private readonly mcpHttpService = createMcpHttpServiceController({
+		app: this.app,
+		pluginId: this.manifest.id,
 		getSettings: () => this.settings,
 		saveSettings: async () => {
 			await this.saveSettings();
 		},
-		getVaultPath: () => getVaultPath(this.app),
-		getPluginDirRealpathOrNull: () => getPluginDirRealpathOrNull(this.app, this.manifest.id),
-		resolveMcpHttpArgs: () =>
-			resolveMcpHttpArgs({
-				settings: this.settings,
-				pluginDirRealpathOrNull: getPluginDirRealpathOrNull(this.app, this.manifest.id),
-			}),
-		getUrl: () => this.getMcpHttpServiceUrl(),
-		onStatusChanged: () => {
+		getMcpUrl: () => this.getMcpHttpServiceUrl(),
+		onMcpStatusChanged: () => {
 			if (!this.mcpStatusBarEl) return;
 			renderMcpStatusBar(this.mcpStatusBarEl, this.getMcpHttpServiceStatusSnapshot());
 		},
+		onPythonStatusChanged: () => undefined,
+		onIndexerSnapshot: () => undefined,
 	});
 
-	private readonly pythonApiService = new PythonApiServiceController({
+	private readonly pythonApiService = createPythonApiServiceController({
+		app: this.app,
+		pluginId: this.manifest.id,
 		getSettings: () => this.settings,
 		saveSettings: async () => {
 			await this.saveSettings();
 		},
-		getVaultPath: () => getVaultPath(this.app),
-		getPluginDirRealpathOrNull: () => getPluginDirRealpathOrNull(this.app, this.manifest.id),
-		resolvePythonApiArgs: () =>
-			resolvePythonApiArgs({
-				settings: this.settings,
-				pluginDirRealpathOrNull: getPluginDirRealpathOrNull(this.app, this.manifest.id),
-			}),
-		getUrl: () => this.getPythonApiServiceUrl(),
-		onStatusChanged: () => {
+		getPythonUrl: () => this.getPythonApiServiceUrl(),
+		onMcpStatusChanged: () => undefined,
+		onPythonStatusChanged: () => {
 			if (!this.pythonStatusBarEl) return;
 			renderPythonStatusBar(this.pythonStatusBarEl, this.getPythonApiServiceStatusSnapshot());
 		},
+		onIndexerSnapshot: () => undefined,
 	});
 
-	private readonly indexer = new IndexerRunner({
+	private readonly indexer = createIndexerRunner({
+		app: this.app,
+		pluginId: this.manifest.id,
 		getSettings: () => this.settings,
 		saveSettings: async () => {
 			await this.saveSettings();
 		},
-		getVaultPath: () => getVaultPath(this.app),
-		getPluginDirRealpathOrNull: () => getPluginDirRealpathOrNull(this.app, this.manifest.id),
-		resolveIndexerArgs: () =>
-			resolveIndexerArgs({
-				settings: this.settings,
-				pluginDirRealpathOrNull: getPluginDirRealpathOrNull(this.app, this.manifest.id),
-			}),
-		onSnapshot: (snapshot) => {
+		onMcpStatusChanged: () => undefined,
+		onPythonStatusChanged: () => undefined,
+		onIndexerSnapshot: (snapshot) => {
 			if (!this.statusBarEl) return;
 			renderIndexerStatusBar(this.statusBarEl, snapshot);
 		},
 	});
 
-	private readonly autoIndex = new AutoIndexScheduler({
+	private readonly autoIndex = createAutoIndexScheduler({
 		getSettings: () => this.settings,
 		isIndexerRunning: () => this.indexer.isRunning(),
 		runIndexer: async (paths) => {
 			await this.indexer.run(paths);
-		},
-		onError: (message) => {
-			showNotice(`AILSS auto-index failed: ${message}`);
 		},
 	});
 
@@ -613,91 +602,4 @@ export default class AilssObsidianPlugin extends Plugin {
 	subscribeIndexerStatus(listener: (snapshot: AilssIndexerStatusSnapshot) => void): () => void {
 		return this.indexer.subscribe(listener);
 	}
-}
-
-function formatPythonRetrieveResult(
-	result: Awaited<ReturnType<typeof requestPythonApiRetrieve>>,
-): string {
-	const lines = [
-		`Query: ${result.query}`,
-		`Mode: ${result.mode}`,
-		`Results: ${result.results.length}`,
-		`Latency: ${Math.round(result.usage.latency_ms)} ms`,
-	];
-
-	if (result.usage.embedding_model) {
-		lines.push(`Embedding model: ${result.usage.embedding_model}`);
-	}
-	if (typeof result.usage.embedding_prompt_tokens === "number") {
-		lines.push(`Embedding prompt tokens: ${result.usage.embedding_prompt_tokens}`);
-	}
-	if (result.warnings.length > 0) {
-		lines.push("", "Warnings:");
-		for (const warning of result.warnings) {
-			lines.push(`- ${warning}`);
-		}
-	}
-
-	if (result.results.length === 0) {
-		lines.push("", "No grounded results.");
-		return lines.join("\n");
-	}
-
-	lines.push("", "Grounded matches:");
-	for (const [index, entry] of result.results.entries()) {
-		lines.push(`${index + 1}. ${entry.path}`);
-		if (entry.title) lines.push(`   Title: ${entry.title}`);
-		if (entry.summary) lines.push(`   Summary: ${entry.summary}`);
-		lines.push(`   Snippet: ${entry.snippet}`);
-		if (entry.evidence.length > 0) {
-			const citationList = entry.evidence.map((chunk) => chunk.chunk_id).join(", ");
-			lines.push(`   Evidence: ${citationList}`);
-		}
-	}
-
-	return lines.join("\n");
-}
-
-function formatPythonAgentResult(result: Awaited<ReturnType<typeof runPythonApiAgent>>): string {
-	const lines = [
-		`Run ID: ${result.run_id}`,
-		`Outcome: ${result.outcome}`,
-		`Retrieval mode: ${result.metrics.retrieval_mode}`,
-		`Selected notes: ${result.metrics.selected_notes}`,
-		`Latency: ${Math.round(result.metrics.latency_ms)} ms`,
-	];
-
-	if (typeof result.metrics.embedding_prompt_tokens === "number") {
-		lines.push(`Embedding prompt tokens: ${result.metrics.embedding_prompt_tokens}`);
-	}
-	if (result.artifact_path) {
-		lines.push(`Artifact: ${result.artifact_path}`);
-	}
-	if (result.answer) {
-		lines.push("", "Answer:", result.answer);
-	}
-	if (result.citations.length > 0) {
-		lines.push("", "Citations:");
-		for (const citation of result.citations) {
-			lines.push(`- ${citation.path}#${citation.chunk_id}`);
-		}
-	}
-	if (result.failure) {
-		lines.push("", "Failure:", `- ${result.failure.code}: ${result.failure.message}`);
-	}
-	if (result.write_actions.length > 0) {
-		lines.push("", "Write actions:");
-		for (const action of result.write_actions) {
-			lines.push(`- ${action.action}: ${action.allowed ? "allowed" : action.reason}`);
-		}
-	}
-	if (result.workflow.length > 0) {
-		lines.push("", "Workflow:");
-		for (const step of result.workflow) {
-			const detailSuffix = step.detail ? ` (${step.detail})` : "";
-			lines.push(`- ${step.name}: ${step.outcome}${detailSuffix}`);
-		}
-	}
-
-	return lines.join("\n");
 }
